requirements.txt
matplotlib==3.3.3
Jinja2==2.11.2
pandas==1.0.0
numpy==1.19.3
datapane==0.10.1
PyYAML==5.4.1
